# Speed up sequential Adds (for sparse bitmaps)