We almost always build bitmaps sequentially and I imagine that's not uncommon. This adds a special case to `(*arrayContainer).add()` which just does a simple append if the value is larger than all others in the container.

```
benchmark                           old ns/op     new ns/op     delta
BenchmarkSequentialAdd-4            38711199      16057751      -58.52%
```

Doesn't seem to hurt random add performance:

```
BenchmarkSetRoaring-4               71.4          70.5          -1.26%
```