@lemire 2015-11-24
Excellent. The Java version does have a similar optimization: 

https://github.com/lemire/RoaringBitmap/blob/master/src/main/java/org/roaringbitmap/Util.java#L202-L205

I cannot tell whether the Go version never had this optimization or whether it was mistakenly removed. In any case, it needs to be there, you are right.